requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
sqlalchemy==2.0.14
psycopg2-binary==2.9.9 
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4  
pydantic==1.10.8
python-multipart==0.0.12